refactor(Store): make `getState()` explicitly
https://github.com/almin/almin/pull/166#discussion_r112360145
Thanks to @kimamula
